requirements.txt
# Notebook-first workflow
jupyter
notebook
ipywidgets

# Core data/scientific stack
numpy
pandas
scipy
matplotlib
seaborn

# Bioinformatics modules (keep)
biopython
pyhmmer
aaindex
esm
openprotein-python
pymol-open-source
nglview

# Retrieval and lightweight app plumbing
requests
httpx
pydantic
python-dotenv
pypdf
tavily-python

# LLM interpretation client (can be swapped later)
openai
